% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abstract_report_template.R
\name{abstract_report}
\alias{abstract_report}
\title{Convert to an HTML vignette.}
\usage{
abstract_report(fig_width = 3, fig_height = 3, dev = "png", css = NULL,
  template = NULL, ...)
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A HTML vignette is a lightweight alternative to \code{\link{html_document}}
suitable for inclusion in packages to be released to CRAN. It reduces the
size of a basic vignette from 100k to around 10k.
}
\details{
Compared to \code{html_document}, it:

\itemize{
  \item never uses retina figures
  \item has a smaller default figure size
  \item uses a custom css stylesheet
 }

See the \href{http://rmarkdown.rstudio.com/package_vignette_format.html}{online
documentation} for additional details on using the \code{html_vignette} format.
}

