% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpharma.r
\name{add_blank_row}
\alias{add_blank_row}
\title{Add a blank row to a data frame}
\usage{
add_blank_row(x, ..., location = "below")
}
\arguments{
\item{x}{The dataframe to add blanks to.}

\item{...}{Column names and non-blank values to assign.}

\item{location}{The location to add the blank row.  Valid values are "above",
"below", and "both".  The default value is "below".}
}
\value{
The input dataset with the blank row added at the specified
location.
}
\description{
The purpose of this function is to add a blank row to the top or bottom
of a dataframe.  Character columns will be set to an empty string.  Numeric
and date columns will be set to an NA. The function allows the user to pass
in values for specified columns.  This feature is useful for setting key
values.  Note that a blank value will be added to factor levels that do not
contain blanks.
}
\examples{
#s <- filter(iris, Species == "setosa")
#b <- add_blank_row(s)
}
