% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpharma.r
\name{add_blank_rows}
\alias{add_blank_rows}
\title{Add blank rows to a data frame after each by-group}
\usage{
add_blank_rows(x, ..., .var_list = NULL)
}
\arguments{
\item{x}{The dataframe to add blanks to.}

\item{...}{Column names for group variables.}

\item{.var_list}{A character vector or list of column names to split by.}
}
\value{
The input dataset with the blank row added after each by-group.
}
\description{
The purpose of this function is to add a blank rows to a dataframe for each
by-group.  Character columns will be set to an empty string.  Numeric
and date columns will be set to an NA. The function allows the user to pass
in column names to group by.  Note that a blank value will be added to factor
levels that do not contain blanks.
}
\examples{
b <- add_blank_rows(iris, Species)
}
