% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\docType{data}
\name{age_group_levels}
\alias{age_group_levels}
\title{Age groups}
\format{
An object of class \code{character} of length 7.
}
\usage{
age_group_levels
}
\description{
Unique values for age group levels
}
\keyword{datasets}
