% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_codebook.R
\name{apply_codebook}
\alias{apply_codebook}
\title{Apply a codebook to a dataset}
\usage{
apply_codebook(codebook, dataset)
}
\arguments{
\item{codebook}{List of class \code{reporter:codebook} or character. A
character vector of length 1 is assumed to be the codebook file
path and is passed to \code{import_codebook()}. No default.}

\item{dataset}{Dataframe. The dataset to which the codebook should
be applied. No default.}
}
\value{
A tibble.
}
\description{
Given a codebook and a dataset this function returns a dataset where the column classes have been modified to fit and factors have been supplied with human readable levels.
}
