% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.R
\name{body_add_flextables}
\alias{body_add_flextables}
\title{Add a list of flextables to a document body}
\usage{
body_add_flextables(
  x,
  flextable_list,
  align = "center",
  pos = "after",
  split = FALSE,
  page_break = TRUE
)
}
\arguments{
\item{x}{officer document to add the flextable to}

\item{flextable_list}{A list of flextables to add to the document}

\item{align}{How to align the table within the document.}

\item{pos}{Where to insert the table within the document.}

\item{split}{Whether to split the table.}

\item{page_break}{Whether to add a page break after each table.}
}
\value{
The original officer document with the tables from the list
added to the document.
}
\description{
Multi-table add function for flextable.  This function avoids the
memory problems associated with adding many multiple tables and
creating large documents.
}
\examples{
# Create sample data
}
\seealso{
[body_add_flextable()]
}
