% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.r
\name{body_add_gg_svg}
\alias{body_add_gg_svg}
\title{Add a ggplot to the document as an SVG graphic}
\usage{
body_add_gg_svg(
  x,
  value,
  width = 6,
  height = 5,
  res = 300,
  style = "Normal",
  ...
)
}
\arguments{
\item{x}{The officer object to add content to.}

\item{value}{The ggplot object.}

\item{width}{The desired width of the graphic.}

\item{height}{The desired height of the graphic.}

\item{res}{The desired resolution of the graphic.}

\item{style}{The style used on the ggplot.}

\item{...}{Other parameters passed to ggsave.}
}
\value{
None
}
\description{
This function will add a SVG format ggplot object to the report body.  Note
that most versions of Word do not support embedded SVG graphics. In that case,
the image will be imbedded as a png.
}
