% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{cnt_pct}
\alias{cnt_pct}
\title{Counts and percent formatting function}
\usage{
cnt_pct(a, b, precision = 0.1, low_threshold = 0.01, low_label = "< 1\%")
}
\arguments{
\item{a}{The vector that contains the count values.}

\item{b}{The vector that contains the percent values.}

\item{precision}{The precision specified for the percent format.}

\item{low_threshold}{The threshold at which the low level applies.}

\item{low_label}{The low label to be used for percents that fall below the
low_threshold.  Default is "< 1\%".}
}
\description{
A function to format counts and percents.
}
