% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook_template.R
\name{create_codebook_template}
\alias{create_codebook_template}
\title{Create a codebook template}
\usage{
create_codebook_template()
}
\value{
A list
}
\description{
Creates an empty codebook object.
}
\examples{
create_codebook_template()
}
