% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_content}
\alias{create_content}
\title{Create content}
\usage{
create_content()
}
\arguments{
\item{codebook}{List of class \code{reporter:codebook}. No default.}

\item{dataset}{Dataframe of class \code{reporter:dataset}. No default.}
}
\description{
Creates report content
}
