% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_reports_summary.R
\name{create_files_table_DT}
\alias{create_files_table_DT}
\title{Create a summary table with DT package}
\usage{
create_files_table_DT(directory)
}
\description{
Create a summary table with DT package
}

