% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_reports_summary.R
\name{create_files_table_DT_source}
\alias{create_files_table_DT_source}
\title{Create a summary table of source files with DT package}
\usage{
create_files_table_DT_source(directory)
}
\description{
Create a summary table of source files with DT package
}

