% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_front_matter.R
\name{create_front_matter}
\alias{create_front_matter}
\title{Create front matter}
\usage{
create_front_matter(
  title = "Report",
  subtitle = NULL,
  author = NULL,
  date = NULL,
  logo = NULL,
  image = NULL,
  write = FALSE
)
}
\arguments{
\item{title}{Character. The report title. Defaults to \code{"Report"}.}

\item{subtitle}{Character. The report subtitle. Defaults to NULL.}

\item{author}{Character. The report author. Defaults to NULL.}

\item{date}{Character. The report date. Defaults to NULL.}

\item{logo}{Character. The path to a valid image file to use as
logo. Defaults to NULL.}

\item{image}{Character. The path to some other front matter
image. Defaults to NULL.}

\item{write}{Logical. If TRUE the front matter is written to disk
and rendered. Defaults to FALSE.}
}
\value{
A character vector with the front matter as markdown.
}
\description{
Creates the first page of the report.
}
\details{
For all arguments, NULL means that the item will not be included.
}
\examples{
## Create front matter with only the title "Report" 
create_front_matter()

## Use a custom title
create_front_matter("My report")

## Create a full front page and write to disk
title <- "My report"
subtitle <- "Made with reporter"
author <- "Just me"
date <- lubridate::today()
logo <- "my-logo.png"
image <- "my-image.png"
create_front_matter(title, subtitle, author, date, logo, image, write = TRUE)
}
