% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_reports_summary.R
\name{create_reports_summary}
\alias{create_reports_summary}
\title{Create a summary file}
\usage{
create_reports_summary(directory = "./", file_name = "toc")
}
\description{
Create a summary file
}

