% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_reports_summary2.R
\name{create_reports_summary2}
\alias{create_reports_summary2}
\title{Create a summary file}
\usage{
create_reports_summary2(directory = "./", type = "html")
}
\arguments{
\item{directory}{}

\item{filename}{}
}
\description{
Create a summary file
}

