% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_reports_summary3.R
\name{create_reports_summary3}
\alias{create_reports_summary3}
\title{Create a summary file}
\usage{
create_reports_summary3(directory, file_name = "toc")
}
\description{
Create a summary file
}

