% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Create summary table}
\usage{
create_summary_table(summary.data, variable)
}
\arguments{
\item{summary.data}{Data frame. The data to be summarised. No
default.}

\item{varible}{A symbol (class \code{name}) or characater. The variable
in \code{summary.data} to be summarised. No default.}
}
\description{
Creates a summary table
}
