% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.r
\name{fit_to_page}
\alias{fit_to_page}
\title{Fit a flextable to a specified width}
\usage{
fit_to_page(ft, page_width = 9)
}
\arguments{
\item{ft}{The flextable to fit.}

\item{page_width}{The width of the available space in inches.}
}
\value{
The flextable returned with column widths resized.
}
\description{
Fit a flex table to take up the amount of space specified by the
\code{pgwidth} parameter.  Default is 9 inches, which will take up
the available body on a landscape report with 1 inch margins.
}
\examples{
# Here is an example
}
