% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullpage.R
\name{fullpage_document}
\alias{fullpage_document}
\title{fabric report}
\usage{
fullpage_document(toc = FALSE, fig_width = 6.5, fig_height = 4,
  fig_retina = 2, fig_caption = FALSE, keep_md = FALSE, smart = TRUE,
  self_contained = TRUE, pandoc_args = NULL, mathjax = "default",
  highlight = "default", ...)
}
\arguments{
\item{toc}{whether to render table of content.}

\item{fig_width, fig_height, fig_retina, fig_caption}{figure dimensions.}

\item{keep_md}{whether to keep markdown.}

\item{smart}{whether to use smart markdown, defaults to \code{TRUE}.}

\item{self_contained}{produce a standalone HTML file with no external dependencies, using data:
URIs to incorporate the contents of linked scripts, stylesheets, images, and videos.
Note that even for self contained documents MathJax is still loaded externally
(this is necessary because of it's size).}

\item{pandoc_args}{additional command line options to pass to pandoc.}

\item{mathjax}{include mathjax. The "default" option uses an https URL from a MathJax CDN.
The "local" option uses a local version of MathJax (which is copied into the output directory).
You can pass an alternate URL or pass NULL to exclude MathJax entirely.}

\item{highlight}{synthax highlighter, defaults to \code{pygment}.}

\item{...}{additional parameters to pass to \code{html_document_base}.}
}
\description{
fabric report
}
