% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gen_groups}
\alias{gen_groups}
\title{gen_groups}
\usage{
gen_groups(tot, group_cnt, last_indices = FALSE)
}
\arguments{
\item{tot}{Total number of items in return vector.}

\item{group_cnt}{Number of items in each group.  This can be a single value
or vector of values.  Will recycle if needed.}

\item{last_indices}{TRUE or FALSE value indicating whether to return the
vector of values, or a vector of row indices where the breaks occur.}
}
\value{
A vector of length \code{tot} broken into groups specified by
\code{group_cnt}.  Groups will be identified by integers from 1 to n.
}
\description{
Creates group values based on a number of items and a vector of item
counts per group.
}
\details{
This function is used to create a vector of group values.  It can use used to
dynamically subset a dataframe based on row counts rather than data values.
The last_indices parameter can be used to return the indices of the last item
in each group.
}
\examples{
gen_groups(10, 3)
#>[1] 1 1 1 2 2 2 3 3 3 4

gen_groups(12, c(3, 2, 5, 2))
#>[1] 1 1 1 2 2 3 3 3 3 3 4 4

gen_groups(10, 3, last_indices = TRUE)
#>[1] 3 6 9 10

gen_groups(12, c(3, 2, 5, 2), last_indices = TRUE)
#>[1] 3 5 10 12
}
