% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpharma.r
\name{\%+\%}
\alias{\%+\%}
\title{Simple Concatenation operator}
\usage{
x \%+\% y
}
\arguments{
\item{x}{The left string to concatenate.}

\item{y}{The right string to concatenate.}
}
\description{
A simple operator for concatenating strings.  This operator
is based on \code{paste0()}, and performs the same function, but with a
more compact syntax.
}
\examples{
"Today is " \%+\% weekdays(Sys.Date()) \%+\% "."
}
\seealso{
[paste0()]
}
