% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_codebook.R
\name{import_codebook}
\alias{import_codebook}
\title{Import codebook}
\usage{
import_codebook(path)
}
\arguments{
\item{path}{Character vector of length 1. The path to the
codebook. Defaults to \code{"./codebook.csv"}, i.e. it looks for a
file called \code{codebook.csv} in the current working directory.}
}
\value{
A list of codebook components indexed according to \code{name}.
}
\description{
Imports a codebook and makes sure that it conforms to the expected
format.
}
