% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{Import dataset}
\usage{
import_dataset(path)
}
\arguments{
\item{path}{Character vector of length 1. The path to the
dataset. Defaults to \code{"./dataset.csv"}, i.e. it looks for a
file called \code{dataset.csv} in the current working directory.}
}
\value{
A tibble.
}
\description{
Imports a dataset
}
\examples{
\dontrun{
import_dataset()
}
}
