% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{materialize_document}
\alias{materialize_document}
\title{fabric report}
\usage{
materialize_document(toc = FALSE, toc_depth = 1, toc_img = NULL,
  toc_side = "left", toc_btn = "red", container = TRUE, fig_width = 6.5,
  fig_height = 4, fig_retina = 2, fig_caption = FALSE, keep_md = FALSE,
  smart = TRUE, self_contained = TRUE, pandoc_args = NULL,
  mathjax = "default", highlight = "default", extra_dependencies = NULL,
  ...)
}
\arguments{
\item{toc}{whether to render table of content.}

\item{toc_depth}{depth of table of content, defaults to \code{2}.}

\item{toc_img}{image to use in table of content.}

\item{toc_side}{side from which table of content should expand.}

\item{toc_btn}{table of content button color.}

\item{container}{whether to use a container, defaults to \code{TRUE}.}

\item{fig_width, fig_height, fig_retina, fig_caption}{figure dimensions.}

\item{keep_md}{whether to keep markdown.}

\item{smart}{whether to use smart markdown, defaults to \code{TRUE}.}

\item{self_contained}{produce a standalone HTML file with no external dependencies, using data:
URIs to incorporate the contents of linked scripts, stylesheets, images, and videos.
Note that even for self contained documents MathJax is still loaded externally
(this is necessary because of it's size).}

\item{pandoc_args}{additional command line options to pass to pandoc.}

\item{mathjax}{include mathjax. The "default" option uses an https URL from a MathJax CDN.
The "local" option uses a local version of MathJax (which is copied into the output directory).
You can pass an alternate URL or pass NULL to exclude MathJax entirely.}

\item{highlight}{synthax highlighter, defaults to \code{pygment}.}

\item{extra_dependencies}{any additional dependencies, see \code{htmltools::htmlDependency}.}

\item{...}{additional parameters to pass to \code{html_document_base}.}
}
\description{
fabric report
}
