% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{options_docx}
\alias{options_docx}
\title{Set options for a docx report}
\usage{
options_docx(x, font_name = "Courier New", font_size = 10)
}
\arguments{
\item{x}{The report spec.}

\item{font_name}{The font name to use on the report.  The specified font
will be used on the entire report.  Valid values are "Courier New", "Arial",
"Calibri", or "Times New Roman".  The default font is "Courier New".}

\item{font_size}{The font size.  Valid value is 10.}
}
\value{
The updated report spec.
}
\description{
This function sets the options for a report of output type docx.
}
\examples{
# Here is an example
}
