% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{options_text}
\alias{options_text}
\title{Set options for a docx report}
\usage{
options_text(x, editor = NULL, cpuom = NULL, lpuom = NULL)
}
\arguments{
\item{x}{The report spec.}

\item{cpuom}{Characters per unit of measure of printed text.
if uom is inches, 
default is 12, which equals a 10pt font.  This value will be used to 
determine how many characters can fit on a line.}

\item{lpuom}{Lines per unit of measure of the printed text. Default for 
inches is 6, which is average for a 10pt font. This value 
will be used to determine the number of lines that can fit on a page.}
}
\value{
The updated report spec.
}
\description{
This function sets the options for a report of output type docx.
}
\examples{
# Here is an example
}
