% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{race_fmt}
\alias{race_fmt}
\title{Function to format race}
\usage{
race_fmt(x)
}
\arguments{
\item{x}{The vector to apply the race format to.}
}
\description{
A vectorized function to format the race variable.
}
