% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{range_fmt}
\alias{range_fmt}
\title{Function to format ranges}
\usage{
range_fmt(a, b)
}
\arguments{
\item{a}{The vector to apply the range format to.}

\item{b}{The vector to apply the range format to.}
}
\description{
A vectorized function to format the range values.
}
