% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.anova.R
\name{report.anova}
\alias{report.anova}
\title{Report the results of ANOVA}
\usage{
report.anova(ordinalna, intervalna, dataset, omegaSq = FALSE)
}
\arguments{
\item{ordinalna}{The name of ordinal variable. One variable can be considered.}

\item{intervalna}{The name of interval/numberic variable(s). A string or a list. Several numeric variables can be considered.}

\item{dataset}{The dataframe with ordinal and numeric variables.}

\item{omegaSq}{Whether to calculate Omega Square (adjusted Etha Square) which is the effect size measure.}
}
\description{
The function reports the results of ANOVA tests with one categorical variable and one or several numeric variables.
}
\examples{
report.anova(ordinalna = "cyl", intervalna = c("hp", "disp", "qsec"), dataset = mtcars)
}
\references{
Kirk, R. E. (1996). Practical significance: A concept whose time has come. Educational and psychological measurement, 56(5), 746-759.
Tunks, T. (1978). The use of omega squared in interpreting statistical significance. Bulletin of the Council for Research in Music Education, 28-34.
}
\author{
Marjan Cugmas
}
