% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.chitest.R
\name{report.chitest}
\alias{report.chitest}
\title{CReport the results of chi-square test}
\usage{
report.chitest(x, y, dataset, simulate.p.value = TRUE, cramer = FALSE, ...)
}
\arguments{
\item{x}{The names of the first set of categorical variables.}

\item{y}{The names of the second set of categorical variables.}

\item{dataset}{The dataframe with categorical variables.}

\item{simulate.p.value}{Wheter to estimate the p-value by simulations (simulate.p.value=TRUE)  or analytically (simulate.p.value=FALSE).}

\item{cramer}{Wheter to calculate the Cramer's V coefficient.}
}
\description{
The function reports the results of several (with several variables) chi-square tests.
}
\note{
One or several categorical variables can be set in x and/or x. When several variables are set, they have to be in a vector format.
}
\examples{
report.chitest(x = c("cyl", "vs", "am"), y = "gear", dataset = mtcars, simulate.p.value = TRUE)
}
\author{
Marjan Cugmas
}
