% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.cortest.R
\name{report.cortest}
\alias{report.cortest}
\title{CReport the results of correlation test}
\usage{
report.cortest(x, y, dataset, method = "pearson")
}
\arguments{
\item{x}{The names of the first set of numerical variables.}

\item{y}{The names of the second set of numerical variables.}

\item{dataset}{The dataframe with numerical variables.}
}
\description{
The function reports the results of several (with several variables) correlation tests.
}
\examples{
report.cortest(x = "drat", y = c("wt", "disp", "hp"), dataset = mtcars)
}
\author{
Marjan Cugmas
}
