% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(entry)
}
\arguments{
\item{entry}{List. The codebook entry for a specific variable with
the variable and if applicable the strata data. No default.}
}
\value{
A character vector formatted as markdown.
}
\description{
Identifies the appropriate report function depending on variable
type
}
