% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_default.R
\name{report_default}
\alias{report_default}
\title{Report default}
\usage{
report_default(entry)
}
\arguments{
\item{entry}{List. The codebook entry for a specific
variable with the variable and if applicable the strata
data. No default.}
}
\value{
A character vector formatted as markdown.
}
\description{
The default report variable function.
}
