% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_variable.R
\name{report_variable}
\alias{report_variable}
\title{Report a variable}
\usage{
report_variable(entry)
}
\arguments{
\item{entry}{List. The codebook entry for a specific variable with
the variable and if applicable the strata data. No default.}
}
\value{
A character vector formatted as markdown.
}
\description{
Identifies the appropriate report function depending on variable
type
}
