\name{reporteR-package}
\alias{reporteR-package}
\alias{reporteR}
\docType{package}
\title{
\packageTitle{reporteR}
}
\description{
\packageDescription{reporteR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{reporteR}
\packageIndices{reporteR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{reporteR}

Maintainer: \packageMaintainer{reporteR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
