% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{sex_fmt}
\alias{sex_fmt}
\title{Function for formatting Sex variable}
\usage{
sex_fmt(x)
}
\arguments{
\item{x}{The vector to apply formatting to.}
}
\description{
A vectorized function to format the sex variable.
}
