% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\docType{data}
\name{sex_levels}
\alias{sex_levels}
\title{Sex levels}
\format{
An object of class \code{character} of length 4.
}
\usage{
sex_levels
}
\description{
The unique levels for the sex variable.
}
\keyword{datasets}
