% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.r
\name{split_df_pages}
\alias{split_df_pages}
\title{split_df_pages}
\usage{
split_df_pages(df, rows, cols, idcols = NULL)
}
\arguments{
\item{df}{A dataframe to split}

\item{rows}{A vector of row counts on which to split.  Will recycle
if needed.}

\item{cols}{A vector of column counts on which to split.  Will recycle
if needed.}

\item{idcols}{A vector of id columns to include on each page.}
}
\value{
A list of dataframes sized according to the specifications
in \code{rows} and \code{cols}
}
\description{
A function to split a dataframe into pages according to vectors rows and cols
}
\examples{
# With row labels and no identity column
}
