% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_variables.R
\name{substitute_variables}
\alias{substitute_variables}
\title{Substitute variables}
\usage{
substitute_variables(template, variables)
}
\arguments{
\item{template}{Character. The name of the template file. No
default.}

\item{variables}{List. The variables. No default.}
}
\value{
A character vector formatted as markdown.
}
\description{
Substitute variables in a R markdown template with the variables
provided. Basically renders the R markdown document using
\code{knitr::knit} and writes it back to a temporary file as markdown.
}
