% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptr.R
\name{table_options}
\alias{table_options}
\title{Illustration of table options}
\description{
The rptr package provides many options for creating statistical
reports.  The below details describe those options.
}
\section{create_report() options}{
{
Here are the options for create_report()
}
}

\section{create_table() options}{
{
Here are the options for create_report()
}
}

\section{define() options}{
{
Here are the options for the define() function.
}
}

