% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.R
\name{theme_normal}
\alias{theme_normal}
\title{Normal theme for flextable.}
\usage{
theme_normal(x, fontname = "Courier New", fontsize = 10)
}
\arguments{
\item{x}{The flextable object.}

\item{fontname}{The name of the font.  Valid values are "Courier New",
"Arial", "Calibri", and "Time New Roman".  Default is "Courier New".}

\item{fontsize}{The size of the font in points.}
}
\value{
The flextable object modified.
}
\description{
A theme for flextable with bold, italic, and row bands turned off.  Just
display the data and don't try to pretty it up.
(Needs more work to eliminate defaults.)
}
