% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.R
\name{use_data_labels}
\alias{use_data_labels}
\title{Use dataframe labels as column headers in flextable}
\usage{
use_data_labels(x)
}
\arguments{
\item{x}{The flextable to modify.}
}
\value{
The flextable returned with header labels assigned.
}
\description{
Assigns the labels associated with a dataframe to the headers in
the flextable.
}
