% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_text.R
\name{write_report_text}
\alias{write_report_text}
\title{Write a text report to the file system}
\usage{
write_report_text(x)
}
\arguments{
\item{x}{The report_spec object to write.}
}
\value{
The report spec.
}
\description{
This function writes a report_spec object to the file system, using the
parameters provided in the object.
}
