% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_disk.R
\name{write_to_disk}
\alias{write_to_disk}
\title{Write to disk}
\usage{
write_to_disk(object, file.name = NULL, keep.temp = FALSE)
}
\arguments{
\item{object}{Character. No default.}

\item{file.name}{Character or NULL. The name of the rendered
file. If NULL the name will be created based on the object
argument. Defaults to NULL.}

\item{keep.temp}{Logical. If TRUE the temporary file used to render
the final document is kept. This is useful for
debugging. Defaults to FALSE.}
}
\value{
Returns \code{object} invisibly.
}
\description{
Write an object to disk and render with \code{package:rmarkdown}.
}
\examples{
front.matter <- create_front_matter()
write_to_disk(front.matter)
}
