% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{add_content}
\alias{add_content}
\title{Add content to a report}
\usage{
add_content(
  x,
  object,
  page_break = TRUE,
  align = "center",
  blank_row = "below"
)
}
\arguments{
\item{x}{The report_spec to append content to.}

\item{object}{The object to append.}

\item{page_break}{Whether to add a page break after the object. 
Valid values are TRUE or FALSE.  You can manipulate the page_break 
parameter to add multiple objects to the same page.}

\item{align}{How to align the content.  Valid values are 'left', 'right',
'center', and 'centre'.}

\item{blank_row}{Whether to put a blank row above or below the content.
Valid values are 'above', 'below', 'both', or 'none'.}
}
\value{
The modified report_spec.
}
\description{
This function adds an object to the report content list.
}
\details{
For a text report, 
valid objects are free text or a table_spec.  Objects will be
appended to the report in order they are added.  By default, a page break
is added after the content.
}
\examples{
library(rptr)
library(magrittr)

# Create temp file path
tmp <- file.path(tempdir(), "mtcars.txt")

# Create the report object
rpt <- create_report(tmp) \%>\% 
  titles("MTCARS Sample Report", align = "left") \%>\% 
  add_content(create_table(mtcars), page_break = FALSE, align = "left") \%>\% 
  add_content(create_text("* NOTE: Car information from 1974.")) 

# Write the report to the file system
write_report(rpt)

# Write report to console
writeLines(readLines(tmp))
}
\seealso{
Other report: 
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{set_margins}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
