% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{add_content}
\alias{add_content}
\title{Add content to a report}
\usage{
add_content(x, object, page_break = TRUE, align = "center")
}
\arguments{
\item{x}{The report_spec to append content to.}

\item{object}{The object to append.}

\item{page_break}{Whether to add a page break after the object. 
Valid values are TRUE or FALSE.  You can manipulate the page_break 
parameter to add multiple objects to the same page.}

\item{align}{How to align the content.  Valid values are 'left', 'right',
'center', and 'centre'.}
}
\value{
The modified report_spec.
}
\description{
This function adds an object to the report content list.
}
\details{
For a text report, 
valid objects are free text or a table_spec.  Objects will be
appended to the report in order they are added.  By default, a page break
is added after the content.
}
\examples{
# Create temp file path
fp <- file.path(tempdir(), "mtcars.txt")

# Create the report object
rpt <- create_report(fp) 

# Add title
rpt <- titles(rpt, "MTCARS sample report")

# Add content 
rpt <- add_content(rpt, create_table(mtcars), page_break = FALSE)
rpt <- add_content(rpt, create_text("* NOTE: Car information from 1971.")) 

# Write the report to the file system
write_report(rpt)

# Write report to console
writeLines(readLines(fp))
}
