% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_style.R
\name{add_style}
\alias{add_style}
\title{Add a style object to a report.}
\usage{
add_style(rpt, style = NULL, theme = NULL)
}
\arguments{
\item{rpt}{The report specification to add a style to.}

\item{style}{A style object which contains style settings to add to the 
report. This parameter is optional.  Default is NULL.  Either add a style
to this parameter, or pass a theme name to the \strong{theme} parameter.}

\item{theme}{A theme name to use for this report.}
}
\description{
This function will add a style object to a report specification.  The 
style may be added either by passing a style object to the "style" parameter,
or by passing a theme name to the "theme" parameter.
}
\details{
Here are some details on the add style function.
}
\section{Themes}{

There are currently seven themes available.  All use Arial 10pt font:
\itemize{
  \item \strong{Plain:} 
  \item \strong{DarkRed:} Arial font, a very light blue document background,
  and table headers and titles in shades of red.
  \item \strong{SeaGreen:} Arial font, a very light blue document background,
  and table headers and titles in shades of green.
  \item \strong{BasicGrey:} Arial font, a very light blue document background,
  and table headers and titles in shades of grey
}
}

\examples{
library(reporter)
library(magrittr)

# Prepare data
dat <- as.data.frame(HairEyeColor)
dat <- dat[dat$Freq >= 10, ]

## Example 1: Use Pre-defined Theme ##

# Create temp file path
tmp1 <- file.path(tempdir(), "HairAndEyes1.html")

# Create table object
tbl <- create_table(dat, borders = "outside") \%>\% 
titles("Hair and Eye Colors with Theme") \%>\% 
column_defaults(width = .6)

# Create report and add theme
rpt <- create_report(tmp1, output_type = "HTML") \%>\% 
       add_content(tbl) \%>\% 
       add_style(theme = "SteelBlue")

# Write out the report        
write_report(rpt)

# Uncomment to View report
# file.show(tmp1)

#' ## Example 2: Create Custom Style ##

# Create temp file path
tmp2 <- file.path(tempdir(), "HairAndEyes2.html")

# Define custom style
sty <- create_style(font_name = "Times",
                    font_size = 10,
                    title_font_size = 12,
                    title_font_bold = TRUE,
                    title_font_color = "Blue",
                    table_header_background = "Blue",
                    table_header_font_bold = TRUE,
                    table_header_font_color = "White",
                    table_body_background = "White",
                    table_body_stripe = "Red")

# Create table object
tbl <- create_table(dat, borders = "outside") \%>\% 
titles("Hair and Eye Colors with Style") \%>\% 
column_defaults(width = .6)

# Create report and add theme
rpt <- create_report(tmp2, output_type = "HTML") \%>\% 
       add_content(tbl) \%>\% 
       add_style(style = sty)

# Write out the report        
write_report(rpt)

# Uncomment to View report
# file.show(tmp2)
}
\seealso{
Other styles: 
\code{\link{create_style}()},
\code{\link{get_theme}()},
\code{\link{print.style_spec}()}
}
\concept{styles}
