% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_style.R
\name{add_style}
\alias{add_style}
\title{Add a style object to a report.}
\usage{
add_style(rpt, style = NULL, theme = NULL)
}
\arguments{
\item{rpt}{The report specification to add a style to.}

\item{style}{A style object which contains style settings to add to the 
report. This parameter is optional.  Default is NULL.  Either add a style
to this parameter, or pass a theme name to the \strong{theme} parameter.}

\item{theme}{A theme name to use for this report.}
}
\description{
This function will add a style object to a report specification.  The 
style may be added either by passing a style object to the "style" parameter,
or by passing a theme name to the "theme" parameter.
}
\details{
Here are some details on the add style function.
}
\section{Themes}{

There are currently seven themes available.  All use Arial 10pt font:
\itemize{
  \item \strong{Plain:} 
  \item \strong{DarkRed:} Arial font, a very light blue document background,
  and table headers and titles in shades of red.
  \item \strong{SeaGreen:} Arial font, a very light blue document background,
  and table headers and titles in shades of green.
  \item \strong{BasicGrey:} Arial font, a very light blue document background,
  and table headers and titles in shades of grey
}
}

\examples{
library(reporter)
}
\seealso{
Other styles: 
\code{\link{create_style}()},
\code{\link{get_theme}()},
\code{\link{print.style_spec}()}
}
\concept{styles}
