% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_content}
\alias{create_content}
\title{Create content}
\usage{
create_content(codebook, dataset, strata = NULL)
}
\arguments{
\item{codebook}{List of class \code{reporter:codebook}. No default.}

\item{dataset}{Dataframe of class \code{reporter:dataset}. No default.}

\item{strata}{Character or NULL. The name of the variable to use
for stratification. The report will not be stratified if
\code{strata} is NULL. Defaults to NULL.}
}
\description{
Creates report content
}
