% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{create_plot}
\alias{create_plot}
\title{Create plot content}
\usage{
create_plot(x, height = NULL, width = NULL)
}
\arguments{
\item{x}{The plot to create.}

\item{height}{The height of the plot in the specified units of measure.}

\item{width}{The width of the plot in the specified units of measure.}
}
\value{
The plot specification.
}
\description{
Function to create a text specification that can be 
added as content to a report.  The function creates an S3 object of class
'text_spec'.  The \code{create_text} function can be used to include 
analysis on a statistical report.
}
\details{
To add plain text to a report, use the \code{create_text} function.  The 
function allows you to set a width and alignment for the text.  The
function will preserve any other formatting you apply to the text.  See
the \code{\link{add_content}} function to control page breaking and 
blanks spaces above or below the text.  

The text specification also accepts titles and footnotes.  See the 
\code{\link{titles}} and \code{\link{footnotes}} functions for further 
details.
}
\examples{
library(rptr)
library(ggplot2)
library(magrittr)

# Create temp file path
tmp <- file.path(tempdir(), "mtcars.pdf")

# Create ggplot
p <- ggplot(mtcars, aes(x=cyl, y=mpg)) + geom_point()

# Create plot object
plt <- create_plot(p, height = 4, width = 8)

rpt <- create_report(tmp, output_type = "PDF") \%>\%
  page_header("Client", "Study: XYZ") \%>\% 
  titles("Figure 1.0", "MTCARS Miles per Cylinder Plot") \%>\%
  set_margins(top = 1, bottom = 1) \%>\%
  add_content(plt) \%>\%
  footnotes("* Motor Trend, 1974") \%>\% 
  page_footer("Time", "Confidential", "Page [pg] of [tpg]")

# Write out report
write_report(rpt)

# Uncomment to view PDF file
# shell.exec(tmp)
}
\seealso{
\code{\link{titles}} to add a title block to the plot,  
\code{\link{footnotes}} to add footnotes, and \code{\link{add_content}} 
to add the plot object to a report.
}
\concept{plot}
