% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{create_report}
\alias{create_report}
\title{Create a report}
\usage{
create_report(file_path = "", output_type = "text", orientation = "landscape")
}
\arguments{
\item{file_path}{The output path of the desired report. Either a full path or
a relative path is acceptable.  This parameter is not required to create the
report_spec object, but will be required to print the report.}

\item{output_type}{The report output type.  Valid values are "text" and 
"docx".  Default is "text".}

\item{orientation}{The page orientation of the desired report.  Valid values
are "landscape" or "portrait".  The default page orientation is "landscape".}
}
\value{
A new report_spec S3 object.
}
\description{
This function is the constructor for the report_spec object.  The report_spec
object contains information needed to construct a report. The object is
defined as an S3 object, and is inherited from a list.
}
\examples{
# Here is an example
rpt <- create_report("output/mtcars.docx", orientation="portrait") 
rpt <- add_content(rpt, create_table(mtcars)) 
#write_report(rpt)
}
\seealso{
\itemize{
  \item \code{\link{titles}} to add titles to the report.
  \item \code{\link{footnotes}} to add footnotes to the report.
  \item \code{\link{page_header}} to add a page header to the report. 
  \item \code{\link{page_footer}} to add a page_footer to the report. 
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{options_text}} to set options for text output.
  \item \code{\link{options_docx}} to set options for docx output.
  \item \code{\link{write_report}} to write the report to the file system.
}
}
