% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{create_report}
\alias{create_report}
\title{Create a report}
\usage{
create_report(
  file_path = "",
  output_type = "text",
  orientation = "landscape",
  uom = "inches",
  paper_size = "letter"
)
}
\arguments{
\item{file_path}{The output path of the desired report. Either a full path or
a relative path is acceptable.  This parameter is not required to create the
report_spec object, but will be required to print the report.}

\item{output_type}{The report output type.  Currently, the only valid value 
is "text".  Default is "text".  Will eventually support "RTF", "PDF", and 
"DOCX".}

\item{orientation}{The page orientation of the desired report.  Valid values
are "landscape" or "portrait".  The default page orientation is "landscape".}

\item{uom}{Specifies the units of measurement.  This setting will 
indicate the units for columns widths, margins, paper size, and other 
measurements. Valid values are "inches" or "cm".  Default value is "inches".}

\item{paper_size}{The expected paper size on which the report may be 
printed.  The \code{paper_size} will determine how much text can fit on
one page.  Valid values are "letter", "legal", "A4", and "RD4".  Default is 
"letter".}
}
\value{
A new report_spec object.
}
\description{
Creates a report shell to which you may add titles, footnotes, content, etc.
}
\details{
This function is the constructor for the report object.  The report
object contains information needed to create a report. The object is
defined as an S3 object, and has a class of 'report_spec'.

The report object holds information concerning report page size, orientation, 
titles, footnotes, page header, page footer, margins, and other options.  
Use the \code{\link{add_content}} function to add content to the report.  
The report may be written to a file using the \code{\link{write_report}} 
function.
}
\section{Report family of functions}{

The report is the primary container for report specifications.  The
following functions add additional specifications to the report object 
initialized with \code{create_report}.
\itemize{
  \item \code{\link{titles}} to add titles to the report.
  \item \code{\link{footnotes}} to add footnotes to the report.
  \item \code{\link{page_header}} to add a page header to the report. 
  \item \code{\link{page_footer}} to add a page_footer to the report. 
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{options_fixed}} to set options for text output.
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{write_report}} to write the report to the file system.
}

The report family of functions are pipe-friendly.  After creating the 
report, you may pipe the object to any of the above functions to append
additional options.
}

\examples{
library(rptr)
library(magrittr)

# Create temp file path
tmp <- file.path(tempdir(), "mtcars.txt")

# Create the report object
rpt <- create_report(tmp, orientation = "portrait") \%>\% 
  titles("MTCARS Sample Report") \%>\% 
  add_content(create_table(mtcars)) 

# Write the report to the file system
write_report(rpt)

# Write the report to the console
writeLines(readLines(tmp))
}
\seealso{
\code{\link{create_table}} and \code{\link{create_text}} functions
to create content for the report.

Other report: 
\code{\link{add_content}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{set_margins}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
