% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{create_report}
\alias{create_report}
\title{Create a report}
\usage{
create_report(
  file_path = "",
  output_type = "TXT",
  orientation = "landscape",
  units = "inches",
  paper_size = "letter",
  missing = ""
)
}
\arguments{
\item{file_path}{The output path of the desired report. Either a full path or
a relative path is acceptable.  This parameter is not required to create the
report_spec object, but will be required to print the report.}

\item{output_type}{The report output type.  Currently, the only valid value 
is "TXT".  Default is "TXT".  Will eventually support "RTF", "PDF", and 
"DOCX".}

\item{orientation}{The page orientation of the desired report.  Valid values
are "landscape" or "portrait".  The default page orientation is "landscape".}

\item{units}{Specifies the units of measurement.  This setting will 
indicate the units for columns widths, margins, paper size, and other 
measurements. Valid values are "inches" or "cm" (centimeters).  
Default value is "inches".}

\item{paper_size}{The expected paper size on which the report may be 
printed.  The \code{paper_size} will determine how much text can fit on
one page.  Valid values are "letter", "legal", "A4", and "RD4".  Default is 
"letter".}

\item{missing}{How to display missing values in the report.  Default is
to replace them with an empty string, which removes them from the report.
To display missing values as is, set the missing parameter to NULL.  To
replace missing values with a character string (such as ".", "-", or "<NA>")
pass the desired character string to the missing parameter.}
}
\value{
A new report_spec object.
}
\description{
Creates a report shell to which you may add titles, footnotes, content, etc.
}
\details{
This function is the constructor for the report object.  The report
object contains information needed to create a report. The object is
defined as an S3 object, and has a class of 'report_spec'.

The report object holds information concerning report page size, orientation, 
titles, footnotes, page header, page footer, margins, and other options.  
Use the \code{\link{add_content}} function to add content to the report.  
The report may be written to a file using the \code{\link{write_report}} 
function.
}
\section{Report family of functions}{

The report is the primary container for report specifications.  The
following functions add additional specifications to the report object 
initialized with \code{create_report}.
\itemize{
  \item \code{\link{titles}} to add titles to the report.
  \item \code{\link{footnotes}} to add footnotes to the report.
  \item \code{\link{page_header}} to add a page header to the report. 
  \item \code{\link{page_footer}} to add a page_footer to the report. 
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{options_fixed}} to set options for text output.
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{write_report}} to write the report to the file system.
}

The report family of functions are pipe-friendly.  After creating the 
report, you may pipe the object to any of the above functions to append
additional options.
}

\section{Notes on PDF output type}{


The PDF output type requires the \strong(rmarkdown) package, a LaTeX renderer 
like \strong(MiKTeX), and the PDF manipulation program \strong(qpdf).  
\strong(MiKTeX) and \strong(qpdf) must be installed separately, and will
not be installed as part of the \strong(rptr) install.  These programs must
furthermore be added to the _path_ environment variable on the system
where \strong(rptr) will run. 

Also note that PDF generation is extremely slow compared to TXT and RTF.  If
you have a very large report, you can improve performance by outputting to
RTF, and then converting to PDF with a document editor or other
conversion program. 

PDF output also supports a limited character set.  Non-ASCII characters
can cause the LaTeX rendering to fail.  Non-ASCII characters can include
special characters, mathematical symbols, and characters from international
languages like Chinese, Japanese, and Korean.

Lastly, note that alignment on some styles of PDF report is not rendered
accurately.  This misalignment is due to LaTeX rendering of plain text
content.  

Future versions of \strong(rptr) will eliminate the LaTeX 
intermediary, and increase performance, broaden the allowed character set, 
and improve alignment.
}

\examples{
library(rptr)
library(magrittr)

# Create temp file path
tmp <- file.path(tempdir(), "mtcars.txt")

# Create the report object
rpt <- create_report(tmp, orientation = "portrait") \%>\% 
  titles("MTCARS Sample Report") \%>\% 
  add_content(create_table(mtcars[1:10, ])) 

# Write the report to the file system
write_report(rpt)

# Write the report to the console
writeLines(readLines(tmp))

#                              MTCARS Sample Report
# 
#   mpg    cyl   disp     hp   drat     wt   qsec     vs     am   gear   carb
#  ----------------------------------------------------------------------------
#    21      6    160    110    3.9   2.62  16.46      0      1      4      4
#    21      6    160    110    3.9  2.875  17.02      0      1      4      4
#  22.8      4    108     93   3.85   2.32  18.61      1      1      4      1
#  21.4      6    258    110   3.08  3.215  19.44      1      0      3      1
#  18.7      8    360    175   3.15   3.44  17.02      0      0      3      2
#  18.1      6    225    105   2.76   3.46  20.22      1      0      3      1
#  14.3      8    360    245   3.21   3.57  15.84      0      0      3      4
#  24.4      4  146.7     62   3.69   3.19     20      1      0      4      2
#  22.8      4  140.8     95   3.92   3.15   22.9      1      0      4      2
#  19.2      6  167.6    123   3.92   3.44   18.3      1      0      4      4
}
\seealso{
\code{\link{create_table}} and \code{\link{create_text}} functions
to create content for the report.

Other report: 
\code{\link{add_content}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{print.report_spec}()},
\code{\link{set_margins}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
