% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{create_report}
\alias{create_report}
\title{Create a report}
\usage{
create_report(
  file_path = "",
  output_type = "text",
  orientation = "landscape",
  uom = "inches",
  paper_size = "letter"
)
}
\arguments{
\item{file_path}{The output path of the desired report. Either a full path or
a relative path is acceptable.  This parameter is not required to create the
report_spec object, but will be required to print the report.}

\item{output_type}{The report output type.  Valid values are "text" and 
"docx".  Default is "text".}

\item{orientation}{The page orientation of the desired report.  Valid values
are "landscape" or "portrait".  The default page orientation is "landscape".}

\item{uom}{Specifies the units of measurement.  This setting will 
indicate the units for columns widths, margins, paper size, and other 
measurements. Valid values are "inches" or "cm".  Default value is "inches".}

\item{paper_size}{The expected paper size on which the report may be 
printed.  The \code{paper_size} will determine how much text can fit on
one page.  Valid values are "letter", "legal", "A4", and "RD4".  Default is 
"letter".}
}
\value{
A new report_spec S3 object.
}
\description{
Creates a report shell to which you may add content.
}
\details{
This function is the constructor for the report_spec object.  The report_spec
object contains information needed to construct a report. The object is
defined as an S3 object, and is inherited from a list.

The report_spec object hold information concerning report page size, font, 
titles, footnotes, page header, page footer, margins, and other options.  
Use the \code{\link{add_content}} function to add content to the report.  
The report may be written to a file using the \code{\link{write_report}} 
function.
}
\examples{
# Create temp file path
fp <- file.path(tempdir(), "mtcars.txt")

# Create the report object
rpt <- create_report(fp) 

# Add title
rpt <- titles(rpt, "MTCARS sample report")

# Add content 
rpt <- add_content(rpt, create_table(mtcars)) 

# Write the report to the file system
write_report(rpt)

# Write report to console
writeLines(readLines(fp))
}
\seealso{
\itemize{
  \item \code{\link{titles}} to add titles to the report.
  \item \code{\link{footnotes}} to add footnotes to the report.
  \item \code{\link{page_header}} to add a page header to the report. 
  \item \code{\link{page_footer}} to add a page_footer to the report. 
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{options_text}} to set options for text output.
  \item \code{\link{add_content}} to add content to the report.
  \item \code{\link{write_report}} to write the report to the file system.
}
}
