% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a report}
\usage{
create_report(codebook, dataset, write = FALSE)
}
\arguments{
\item{codebook}{List of class \code{reporter:codebook} or character. A
character vector of length 1 is assumed to be the codebook file
path and is passed to \code{import_codebook()}. No default.}

\item{dataset}{Dataframe of class \code{reporter:dataset} or
character. A character vector of length 1 is assumed to be the
dataset file path and is passed to \code{import_dataset()}. No
default.}

\item{write}{Logical. If TRUE the report is written to disk using
\code{write_report()}. If FALSE the report is only
returned. Defaults to FALSE.}
}
\value{
A report, which is a list of report components that can be
written to disk using \code{write_report()} or using
\code{create_report(.., write = TRUE)}
}
\description{
Creates a data report using a codebook and dataset. The codebook
and dataset have to be aligned so that the codebook describes the
content of the dataset.
}
\examples{
create_report()
}
