% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{create_table}
\alias{create_table}
\title{A function to create a table_spec object}
\usage{
create_table(
  x,
  n_format = upcase_parens,
  page_var = NULL,
  show_cols = "all",
  first_row_blank = FALSE,
  headerless = FALSE
)
}
\arguments{
\item{x}{The data frame to create a table spec for.}

\item{n_format}{The format function to apply to the header n label.}

\item{page_var}{A variable in the data frame to use for a page variable.}

\item{show_cols}{Whether to show all column by default.  Valid values are
"all", "none", or a vector of column names.  "all" means show all columns 
by default, unless overridden by the column definitions.  
"none" means don't show any 
columns unless specified in the column definitions.  If a vector of column
names is supplied, those columns will be shown in the report, whether
or not a definition is supplied.}

\item{first_row_blank}{Whether to place a blank row under the table header.}

\item{headerless}{Whether to create a headerless table.  Default is FALSE.}

\item{align}{Aligns the table on the page.  Valid values are "left", 
"right", and "center".  Default value is "center".}
}
\description{
A function to create a table_spec object
}
