% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_text.R, R/text_spec.R
\name{create_text}
\alias{create_text}
\title{Create a text specification}
\usage{
create_text(txt, width = NULL, align = "left")

create_text(txt, width = NULL, align = "left")
}
\arguments{
\item{txt}{The text to create.}

\item{width}{The width of the text in the specified units of measure.  If 
no width is specified, the full page width will be used.}

\item{align}{How to align the text within the text area.  Valid values
are 'left', 'right', 'center', or 'centre'.}
}
\value{
The text specification.

The text specification.
}
\description{
Function to create a text specification that can be 
added as content to a report.

Function to create a text specification that can be 
added as content to a report.
}
\details{
To add plain text to a report, use the \code{create_text} function.  The 
function allows you to set a width and justification for the text.  The
function will preserve any other formatting you apply to the text.  See
the \code{\link{add_content}} function to control page breaking and 
blanks spaces above or below the text.
}
\concept{text}
