% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_text.R
\name{create_text}
\alias{create_text}
\title{Create a text specification}
\usage{
create_text(txt, width = NULL, align = "left", blank_row = "below")
}
\arguments{
\item{txt}{The text to create.}

\item{width}{The width of the text in the specified units of measure.}

\item{align}{How to align the text within the text area.  Valid values
are 'left', 'right', 'center', or 'centre'.}

\item{blank_row}{Whether to create blanks above or below the object.  Valid
value are 'above', 'below', 'both', or 'none'.}
}
\value{
The text specification.
}
\description{
Function to create a text specification that can be 
added as content to a report.
}
