% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_text.R
\name{create_text}
\alias{create_text}
\title{Create a text specification}
\usage{
create_text(txt, blanks = "after", width = NULL)
}
\arguments{
\item{txt}{The text to create.}

\item{blanks}{Whether to create blanks before or after the object.  Valid
value are 'before', 'after', or 'none'.}

\item{width}{The width of the text in the specified units of measure.}
}
\value{
The text specification.
}
\description{
Function to create a text specification that can be 
added as content to a report.
}
