% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{define}
\alias{define}
\title{Defines a column specification}
\usage{
define(
  x,
  var,
  label = NULL,
  format = NULL,
  align = NULL,
  label_align = NULL,
  width = NULL,
  visible = TRUE,
  n = NULL,
  blank_after = FALSE,
  dedupe = FALSE,
  id_var = FALSE,
  page_wrap = FALSE,
  indent = NULL,
  label_row = FALSE
)
}
\arguments{
\item{x}{The table spec.}

\item{var}{The variable to define a column for.}

\item{label}{The label to use for the column header.}

\item{format}{The format to use for the column data.  The format can 
be a string format, a formatting function, or a format object from the 
\strong{fmtr} package.}

\item{align}{The column alignment.  Value values are "left", "right", and
"center".}

\item{label_align}{How to align the header labels for this column.
Value values are "left", "right", and "center".}

\item{width}{The width of the column in inches.}

\item{visible}{Whether or not the column should be visible on the report.}

\item{n}{The n value to place in the n header label.}

\item{blank_after}{Whether to place a blank row after unique values of this
variable.}

\item{dedupe}{Whether to dedupe the values for this variable.  Variables
that are deduped only show the value on the first row in a group.}

\item{id_var}{Whether this variable should be considered an ID variable.
ID variables are retained on each page when the page is wrapped.}

\item{page_wrap}{Force a page wrap on this variable.  A page wrap is a vertical
page break necessary when the table is too wide to fit on a single page.
The excess variables will be wrapped to the next page.}

\item{indent}{How much to indent the column values.  Parameter takes a 
numeric value that will be interpreted according to the 'uom' 
(Unit Of Measure) setting on the report.}

\item{label_row}{Whether the values of the variable should be used to
create a label row.  Valid values are TRUE or FALSE.  Default is FALSE.
If label_row is set to TRUE, the dedupe parameter will also be set to TRUE.}
}
\description{
Defines a column specification
}
