% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{define}
\alias{define}
\title{Defines a column specification}
\usage{
define(
  x,
  var,
  label = NULL,
  col_type = NULL,
  align = NULL,
  label_align = NULL,
  width = NULL,
  visible = TRUE,
  n = NULL,
  blank_after = FALSE,
  dedupe = FALSE,
  id_var = FALSE
)
}
\arguments{
\item{x}{The table spec.}

\item{var}{The variable to define a column for.}

\item{label}{The label to use for the column header.}

\item{col_type}{The column type.}

\item{align}{The column alignment.  Value values are "left", "right", and
"center".}

\item{label_align}{How to align the header labels for this column.
Value values are "left", "right", and "center".}

\item{width}{The width of the column in inches.}

\item{visible}{Whether or not the column should be visible on the report.}

\item{n}{The n value to place in the n header label.}

\item{blank_after}{Whether to place a blank row after unique values of this
variable.}

\item{dedupe}{Whether to dedupe the values for this variable.  Variables
that are deduped only show the value on the first row in a group.}

\item{id_var}{Whether this variable should be considered an ID variable.
ID variables are retained on each page when the page is wrapped.}
}
\description{
Defines a column specification
}
