% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{footnotes}
\alias{footnotes}
\title{Add footnotes to the report}
\usage{
footnotes(x, ..., align = "left", blank_row = "above")
}
\arguments{
\item{x}{The object to assign footnotes to.}

\item{...}{A set of footnotes strings.}

\item{align}{The position to align the titles.  Valid values are: "left",
"right", "center", or "centre".}

\item{blank_row}{Whether to print a blank row above or below the footnote.
Valid values are 'above', 'below', 'both', or 'none'.  Default is 'above'.}
}
\value{
The modified report.
}
\description{
This function adds one or more footnotes to the report.  The footnotes will
be added to the page template, and thus appear on each page of the report.
}
\details{
The footnotes function accepts a set of strings of the desired footnote text.
The footnotes may be aligned center, left or right using the align parameter.
The footnotes may be located in the footer, page body, or table according to
the location parameter. The user is responsible for adding desired symbols.
Footnote symbols will not be generated automatically.
}
\examples{
library(rptr)
library(magrittr)

# Create a temporary file
tmp <- file.path(tempdir(), "mtcars.txt")

# Define report
rpt <- create_report(tmp, orientation="portrait") \%>\%
  titles("Table 1.0", "MTCARS Sample Report") \%>\% 
  add_content(create_table(mtcars)) \%>\%
  footnotes("* From Motor Trend Magazine, 1974")  

# Write the report
write_report(rpt)

# Display in console
writeLines(readLines(tmp))
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{set_margins}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
