% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{footnotes}
\alias{footnotes}
\title{Add footnotes to the report}
\usage{
footnotes(x, ..., location = "footer", align = "left")
}
\arguments{
\item{x}{The report specification object.}

\item{...}{A set of footnotes strings.}

\item{location}{The location to place the footnotes in the report.
Valid values are "header", "body", or "table".}

\item{align}{The position to align the titles.  Valid values are: "left",
"right", or "center".}
}
\description{
This function adds one or more footnotes to the report.  The footnotes will
be added to the page template, and thus appear on each page of the report.
}
\details{
The footnotes function accepts a set of strings of the desired footnote text.
The footnotes may be aligned center, left or right using the align parameter.
The footnotes may be located in the footer, page body, or table according to
the location parameter. The user is responsible for adding desired symbols.
Footnote symbols will not be generated automatically.
}
\examples{
# Here are some examples.
}
