% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullpage.R
\name{fullpage_document}
\alias{fullpage_document}
\title{fabric report}
\usage{
fullpage_document(toc = FALSE, center = TRUE, background = NULL,
  navigation = FALSE, continuousVertical = TRUE,
  continuousHorizontal = TRUE, navigationPosition = "right",
  slidesNavigation = FALSE, fig_width = 6.5, fig_height = 4,
  fig_retina = if (!fig_caption) 2, fig_caption = FALSE, keep_md = FALSE,
  smart = TRUE, self_contained = TRUE, pandoc_args = NULL,
  highlight = "default", mathjax = "default", includes = NULL,
  css = NULL, extra_dependencies = NULL, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{center}{whether to center text horizontally and vertically.}

\item{background}{background colors, see details.}

\item{navigation}{whether to show navigation bar as small circles.}

\item{continuousVertical}{efines whether scrolling down in the last section or should scroll d
own to the first one and if scrolling up in the first section should scroll up to the last one.}

\item{continuousHorizontal}{Defines whether sliding right in the last slide should slide right
to the first one or not, and if scrolling left in the first slide should slide left to the last one or not.}

\item{navigationPosition}{Defines which position the navigation bar will be shown (if using one).
Takes \code{left} or \code{right}.}

\item{slidesNavigation}{If set to \code{TRUE} it will show a navigation bar made up of small
circles for each landscape slider on the site.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays (defaults to 2, which
currently works for all widely used retina displays). Set to \code{NULL} to
prevent retina scaling. Note that this will always be \code{NULL} when
\code{keep_md} is specified (this is because \code{fig_retina} relies on
outputting HTML directly into the markdown document).}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
"haddock", and "textmate". Pass \code{NULL} to prevent syntax highlighting.}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from a
MathJax CDN. The "local" option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{css}{One or more css files to include}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{...}{additional parameters to pass to \code{html_document_base}.}
}
\description{
fabric report
}
