% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.r
\name{header_image}
\alias{header_image}
\title{Add an image in page header}
\usage{
header_image(x, image_path, height, width, align = "left")
}
\arguments{
\item{x}{The report object.}

\item{image_path}{The path of image}

\item{height}{Height of the image}

\item{width}{Width of the image}

\item{align}{Alignment of the image. Should be "left", "right", "center", or "centre". 
Default is "left"}
}
\value{
The modified report specification.
}
\description{
This function adds an image in page header to the report.  The image will appear
at the top of each page of the report. It only works when page_header is used.
}
\details{
This function adds an image in page header to the report.  The image will appear
at the top of each page of the report. It only works when page_header is used.
If the text is also defined in page header, the text will be ignored.
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footer_image}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_by}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{print.report_spec}()},
\code{\link{set_margins}()},
\code{\link{title_header}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
