% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lowcase_parens}
\alias{lowcase_parens}
\alias{upcase_parens}
\alias{lowcase_n}
\alias{upcase_n}
\title{Functions to format the population label}
\usage{
lowcase_parens(x)

upcase_parens(x)

lowcase_n(x)

upcase_n(x)
}
\arguments{
\item{x}{Population count}
}
\description{
Functions to format the population label
}
