% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{materialize_document}
\alias{materialize_document}
\title{fabric report}
\usage{
materialize_document(toc = FALSE, toc_depth = 1, toc_bg = NULL,
  toc_logo = NULL, toc_color = "black-text", toc_side = "left",
  toc_btn = "red", toc_fixed = FALSE, banner = NULL, container = TRUE,
  fig_width = 6.5, fig_height = 4, fig_retina = 2, fig_caption = FALSE,
  keep_md = FALSE, smart = TRUE, self_contained = TRUE,
  pandoc_args = NULL, mathjax = "default", highlight = "default",
  extra_dependencies = NULL, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{toc_bg}{table of content background, path to an image.}

\item{toc_logo}{image to use in table of content.}

\item{toc_color}{table of content text color, a material class.}

\item{toc_side}{side from which table of content should expand.}

\item{toc_btn}{table of content button color.}

\item{toc_fixed}{whether to use a fixed table of content, defaults to \code{FALSE}.}

\item{banner}{banner image, optional.}

\item{container}{whether to use a container, defaults to \code{TRUE}.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays (defaults to 2, which
currently works for all widely used retina displays). Set to \code{NULL} to
prevent retina scaling. Note that this will always be \code{NULL} when
\code{keep_md} is specified (this is because \code{fig_retina} relies on
outputting HTML directly into the markdown document).}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from a
MathJax CDN. The "local" option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
"haddock", and "textmate". Pass \code{NULL} to prevent syntax highlighting.}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{...}{additional parameters to pass to \code{html_document_base}.}
}
\description{
fabric report
}
