% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptr.R
\name{NotesOnPDF}
\alias{NotesOnPDF}
\title{Notes on PDF output type}
\description{
The PDF output type has some limitations not associated with 
TXT and RTF output.  PDF has special installation requirements and package
dependencies.
}
\details{
The PDF output type requires the \strong{rmarkdown} package and a 
LaTeX renderer like \strong{MiKTeX}.  
\strong{MiKTeX} must be installed separately, and will
not be installed as part of the \strong{rptr} install.  This program must
furthermore be added to the \code{path} environment variable on the system
where \strong{rptr} will run, so that \strong{rptr} can find it.

PDF output also supports a limited character set.  Some Unicode characters
can cause the LaTeX rendering to fail.  Common Unicode characters like 
mathematical operators and Greek letters are supported.  However, 
other special characters and characters from Asian languages may not 
be supported, depending on the operating system and installed operating 
system languages. 

Future versions of \strong{rptr} will eliminate the LaTeX 
intermediary and broaden the allowed character set.
}
