% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{NotesOnPDF}
\alias{NotesOnPDF}
\title{Notes on PDF output type}
\description{
PDF output supports a limited character set.  Some Unicode characters
can cause the PDF rendering to fail.  All Latin characters and other 
common characters like 
mathematical operators and Greek letters are supported.  However, 
other special characters and characters from Asian languages may not 
be supported, depending on the operating system and installed operating 
system languages. If a character is not supported, it will be replaced
with a question mark (?).

As of \strong{reporter} v1.1.3, the package will 
generate PDF files directly, without using \strong{rmarkdown} as an 
intermediary. This technique eliminates dependencies, makes the package 
easier to install, and greatly improves performance.  It is recommended
to use the most recent version of the package if possible.

For versions of \strong{reporter} prior to v1.1.3, the package has
additional dependencies and installation requirements.  See details below.
}
\details{
For \strong{reporter} v1.1.2 and below, 
PDF output type requires the \strong{rmarkdown} package and a 
LaTeX renderer like \strong{MiKTeX}.  
\strong{MiKTeX} must be installed separately, and will
not be installed as part of the \strong{reporter} install.  This program must
furthermore be added to the \code{path} environment variable on the system
where \strong{reporter} will run, so that \strong{reporter} can find it.
}
\keyword{internal}
