% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{options_fixed}
\alias{options_fixed}
\title{Set options for a report with a fixed width font}
\usage{
options_fixed(x, editor = NULL, cpuom = NULL, lpuom = NULL)
}
\arguments{
\item{x}{The report spec.}

\item{editor}{The expected text editor to use for printing.  Assigning
this parameter can reduce the number of other other setting you need
to specify to get a properly printed report.  Valid values are 'notepad',
'word', 'wordpad', 'notepad++', and 'editplus'.}

\item{cpuom}{Characters per unit of measure of printed text.
if uom is inches, 
default is 12, which equals a 10pt font.  This value will be used to 
determine how many characters can fit on a line.}

\item{lpuom}{Lines per unit of measure of the printed text. Default for 
inches is 6, which is average for a 10pt font. This value 
will be used to determine the number of lines that can fit on a page.}
}
\value{
The updated report spec.
}
\description{
This function sets the options for a report of output type 'text'
}
\examples{
# Here is an example
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{set_margins}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
