% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{options_text}
\alias{options_text}
\title{Set options for a docx report}
\usage{
options_text(x, cpi = 12)
}
\arguments{
\item{x}{The report spec.}

\item{cpi}{Characters per inch.  Valid values are between 8 and 14.  Default
value is 12.}
}
\value{
The updated report spec.
}
\description{
This function sets the options for a report of output type docx.
}
\examples{
# Here is an example
}
