% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{page_by}
\alias{page_by}
\title{Adds a page by variable}
\usage{
page_by(x, var, label = NULL, align = "left", blank_row = "below")
}
\arguments{
\item{x}{The report specification to assign the page by to.}

\item{var}{The page by variable.  There can be only one page by per report, 
and one page by variable.  The page by can be passed either quoted or 
unquoted.}

\item{label}{A label to be used as a prefix to the page by variable value.
By default, the label will be assigned to the variable name.  Alternatively,
you may specify a string value to use for the label.}

\item{align}{How to align the page by.  Default value is 'left'. Valid
values are 'left', 'right', 'center', or 'centre'.}

\item{blank_row}{Indicates whether a blank row is desired above or below
the page by.  Default value is 'none'.  Valid values are 'above', 'below',
'both', or 'none'.}
}
\description{
This function adds a page by variable to the report or table.  
The page by will appear above the column headers.  The page by can be
added to either the report or the table.
}
\details{
Only one page by is allowed per report or table.  The page by will appear 
on all pages of the object.  The page by may be aligned on the 
left, right, or center. Use the \code{align} parameter to specify the 
alignment.
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{print.report_spec}()},
\code{\link{set_margins}()},
\code{\link{title_header}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
