% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.r
\name{page_footer}
\alias{page_footer}
\title{Add a page footer to the report}
\usage{
page_footer(x, left = "", right = "", center = "")
}
\arguments{
\item{x}{The report spec object.}

\item{left}{The left page footer text.  May be a single string or a vector
of strings.}

\item{right}{The right page footer text.  May be a single string or a vector
of strings.}

\item{center}{The center page footer text.  May be a single string or a
vector of strings.}
}
\description{
This function adds a page footer to the report.  The page footer will appear
on each page of the report.
}
\details{
The page footer may contain text on the left, right, or center.
Use the appropriate parameters to specify the desired text.  The page footer
may also contain footnotes for the report, if the location of the
the footnotes has been specified as "footer".  See \code{footnotes()}
function for additional details.
}
\examples{
# Here is an example.
}
