% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{page_header}
\alias{page_header}
\title{Add a page header to the report}
\usage{
page_header(x, left = "", right = "", blank_row = "none")
}
\arguments{
\item{x}{The report spec object.}

\item{left}{The left page header text.  May be a single string or a vector
of strings.}

\item{right}{The right page header text.  May be a single string or a vector
of strings.}

\item{blank_row}{Whether to create a blank row below the page header.
Valid values are 'below' and 'none'.  Default is 'none'.}
}
\value{
The modified report specification.
}
\description{
This function adds a page header to the report.  The page header will appear
on each page of the report.
}
\details{
The page header may contain text on the left or right. Use the appropriate
parameters to specify the desired text.  Only one page header is allowed
on a report. The page header will be repeated on every page of the report.
Multiple text values for each side
may be specified as a vector of strings.
}
\examples{
library(rptr)
library(magrittr)

# Create temp file path
tmp <- file.path(tempdir(), "mtcars.txt")

# Create the report object
rpt <- create_report(tmp, orientation = "portrait") \%>\% 
  page_header("Client: Motor Trend", "Study: Cars") \%>\% 
  titles("MTCARS Sample Report") \%>\% 
  add_content(create_table(mtcars)) 

# Write the report to the file system
write_report(rpt)

# Write report to console
writeLines(readLines(tmp))
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{set_margins}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
