% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{page_header}
\alias{page_header}
\title{Add a page header to the report}
\usage{
page_header(x, left = "", right = "", blank_row = "none")
}
\arguments{
\item{x}{The report spec object.}

\item{left}{The left page header text.  May be a single string or a vector
of strings.}

\item{right}{The right page header text.  May be a single string or a vector
of strings.}

\item{blank_row}{Whether to create a blank row below the page header.
Valid values are 'below' and 'none'.  Default is 'none'.}
}
\value{
The modified report specification.
}
\description{
This function adds a page header to the report.  The page header will appear
on each page of the report.
}
\details{
The page header may contain text on the left or right. Use the appropriate
parameters to specify the desired text.  Multiple text values for each side
may be specified as a vector of strings.
}
\examples{
# Here is an example
}
