% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{print.plot_spec}
\alias{print.plot_spec}
\title{Prints the plot spec}
\usage{
\method{print}{plot_spec}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The plot spec.}

\item{...}{Additional parameters to pass to the underlying print function.}

\item{verbose}{Whether to print in verbose form.  Default is FALSE.}
}
\value{
The plot spec, invisibly.
}
\description{
A function to print the plot spec.
The \strong{print} function will print the plot spec in summary 
form.  To view all parameters, set the \code{verbose} parameter to TRUE.
}
\examples{
txt <- create_text("Lorem ipsum dolor sit amet, consectetur...")
print(txt)

# A text specification:
# - text: data.frame 'mtcars' 32 rows 11 cols
}
\seealso{
\code{\link{create_plot}} function to create a plot specification.

Other plot: 
\code{\link{create_plot}()}
}
\concept{plot}
