% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{print.report_spec}
\alias{print.report_spec}
\title{Print the report parameters.}
\usage{
\method{print}{report_spec}(x, ..., full = FALSE)
}
\arguments{
\item{x}{The report spec to print.}

\item{...}{Generic arguments}

\item{full}{Whether to print the spec or an abbreviated version.}
}
\description{
A function to print the report parameters.  This function is an S3 generic
function.  The class printed is "report_spec".
}
\examples{
# Here is an example
}
