% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{print.table_spec}
\alias{print.table_spec}
\title{Prints the table spec}
\usage{
\method{print}{table_spec}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The table spec.}

\item{...}{Additional parameters to pass to the underlying print function.}

\item{verbose}{Whether to print in verbose form, which is similar to 
a list.  Default is FALSE, which prints in summary form.}
}
\value{
The table spec, invisibly.
}
\description{
A function to print the table spec.
The \strong{print} function will print the table spec in summary 
form.  To view all parameters, set the \code{verbose} parameter to TRUE.
}
\examples{
tbl <- create_table(mtcars)
print(tbl)

# A table specification:
# - data: data.frame 'mtcars' 32 rows 11 cols
}
\seealso{
\code{\link{create_table}} function to create a table specification.

Other table: 
\code{\link{column_defaults}()},
\code{\link{create_table}()},
\code{\link{define}()},
\code{\link{spanning_header}()},
\code{\link{stub}()}
}
\concept{table}
