% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{print.table_spec}
\alias{print.table_spec}
\title{Prints the table spec}
\usage{
\method{print}{table_spec}(x, ...)
}
\arguments{
\item{x}{The table spec.}

\item{...}{Additional parameters.}
}
\value{
The table spec, invisibly.
}
\description{
Prints the table spec
}
