% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{print.table_spec}
\alias{print.table_spec}
\title{Prints the table spec}
\usage{
\method{print}{table_spec}(x, ...)
}
\arguments{
\item{x}{The table spec.}

\item{...}{Additional parameters to pass to the underlying print function.}
}
\value{
The table spec, invisibly.
}
\description{
A function to print the table spec.
The \strong{print} function will print the table spec primarily in list 
format.  The exception to list format is the data parameter, which will
be printed in data frame/tibble format.
}
\examples{
tbl <- create_table(mtcars)
print(tbl)

# $data
# mpg cyl  disp  hp drat    wt  qsec vs am gear carb
# Mazda RX4         21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
# Mazda RX4 Wag     21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
# Datsun 710        22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
# Hornet 4 Drive    21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
# Hornet Sportabout 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
# Valiant           18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
# Duster 360        14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
# Merc 240D         24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
# Merc 230          22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
# Merc 280          19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
# [ reached 'max' / getOption("max.print") -- omitted 22 rows ]
# 
# $n_format
# function(x) {
#   
#   ret <- paste0("\n(N=", x, ")")
#   
#   return(ret)
#   
# }
# <environment: namespace:rptr>
#   
#   $col_defs
# list()
# 
# $col_spans
# list()
# 
# $show_cols
# [1] "all"
# 
# $first_row_blank
# [1] FALSE
# 
# $headerless
# [1] FALSE
}
\seealso{
\code{\link{create_table}} function to create a table specification.

Other table: 
\code{\link{column_defaults}()},
\code{\link{create_table}()},
\code{\link{define}()},
\code{\link{spanning_header}()},
\code{\link{stub}()}
}
\concept{table}
