% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.r
\name{report_options}
\alias{report_options}
\title{Set options for the report}
\usage{
report_options(
  x,
  allow_code = FALSE,
  line_break = TRUE,
  page_wrap = TRUE,
  auto_page = TRUE,
  title_block = "table"
)
}
\arguments{
\item{x}{The report spec object.}

\item{allow_code}{Whether to insert code into the report. Now it's only for
RTF code insertion. Default is FALSE. When it's TRUE, users can insert RTF code
to the data, titles, or footnotes. If special characters are passed by the
common package, please double the braces.}

\item{line_break}{Whether to use auto line breaking. Default is TRUE. This is
only for RTF, DOCX, and HTML.}

\item{page_wrap}{Whether to use auto page wrapping. Default is TRUE. If it has
been set in \code{\link{create_table}}, then it follows table's setting. Please
note that turning off page wrapping might make the report out of margin.}

\item{auto_page}{Whether to use auto pagination. Default is TRUE. If it has
been set in \code{\link{create_table}}, then it follows table's setting. 
When `page_break` variable is set in \code{\link{define}}, users can set 
`auto_page` as FALSE to let reporter follow the paging variable without any
auto paging. Please note that turning off auto pagination might cause overflow.}

\item{title_block}{To decide whether the structure of titles is "table" or 
"paragraph". Default is "table". It's only for RTF.}
}
\value{
The report_spec with option settings.
}
\description{
This function is to control the report for auto line breaking, auto pagination,
auto page wrapping. For RTF, it provides options for allowing users to insert
RTF code and deciding the structure of titles.
}
\details{
This function is to control the report for auto line breaking, auto pagination,
auto page wrapping. For RTF, it provides options for allowing users to insert
RTF code and deciding the structure of titles.
}
\examples{
library(reporter)
library(magrittr)

# Create a temporary file
tmp <- file.path(tempdir(), "bod.rtf")

# Define table
tbl <- create_table(BOD, width = 2.5) \%>\% 
  titles("Table 3.6", "BOD¹ Sample Report") \%>\% 
  define(Time, format = "Day \%s", align = "left") \%>\% 
  define(demand, format = "\%2.1f mg/l", label = "Demand") \%>\% 
  footnotes("¹ Biochemical Oxygen Demand")
       
# Define report #1 - No blank margins
rpt <- create_report(tmp, orientation="portrait", output_type = "RTF",
                     font = "Arial", font_size = 9) \%>\%
  add_content(tbl) \%>\%
  report_options(allow_code = TRUE, line_break = FALSE, 
                 auto_page = FALSE,page_wrap = FALSE, 
                 title_block = "paragraph")

# Write the report
write_report(rpt)
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footer_image}()},
\code{\link{footnotes}()},
\code{\link{header_image}()},
\code{\link{options_fixed}()},
\code{\link{page_by}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{print.report_spec}()},
\code{\link{set_margins}()},
\code{\link{title_header}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
