% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporteR.R
\docType{package}
\name{reporteR-package}
\alias{reporteR}
\alias{reporteR-package}
\title{reporteR - a package for managing reproducible reports}
\description{
reporteR - a package for managing reproducible reports
}
\author{
Piotr Sobczyk
}

