% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\docType{package}
\name{reporter}
\alias{reporter}
\title{reporter: A package for creating printable reports}
\description{
The reporter package creates tabular reports in ASCII text and
docx format
}
\section{reporter functions}{

\code{\link{create_report}}
\code{\link{create_table}}
\code{\link{write_report}}
}

