% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptr.R
\docType{package}
\name{rptr}
\alias{rptr}
\title{rptr: A package for creating statistical reports}
\description{
The rptr package creates tabular reports in text output format.
    Will soon support RTF, PDF, and DOCX format.
}
\section{rptr functions}{

\itemize{
  \item \code{\link{create_report}}
  \item \code{\link{create_table}}
  \item \code{\link{write_report}}
}
}

