% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{set_margins}
\alias{set_margins}
\title{Set page margins}
\usage{
set_margins(x, top = NULL, bottom = NULL, left = NULL, right = NULL)
}
\arguments{
\item{x}{The report spec object.}

\item{top}{The top margin.}

\item{bottom}{The bottom margin.}

\item{left}{The left margin.}

\item{right}{The right margin.}
}
\value{
The report_spec with margins set as desired.
}
\description{
Set the page margins on the report spec object.
}
\details{
The margins will be used for the entire report.  Units for the margins
are specified using the \strong{uom} parameter on the 
\code{\link{create_report}} function.  Available units are 'inches' and 'cm'.
When the unit of measure is inches, default margins are 1 inch on left and 
right, and .5 inches on top and bottom.  When the unit of measure is 
centimeters, default margins are 2.54 cm on left and right, and 1.27 cm
on top and bottom.
}
\examples{
# Here is an example
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
