% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{set_margins}
\alias{set_margins}
\title{Set page margins}
\usage{
set_margins(
  x,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  min_margin = NULL,
  blank_margins = FALSE
)
}
\arguments{
\item{x}{The report spec object.}

\item{top}{The top margin.}

\item{bottom}{The bottom margin.}

\item{left}{The left margin.}

\item{right}{The right margin.}

\item{min_margin}{The printer minimum margin.  When the units of measure
is set to centimeters, this parameter defaults to 1.  When the units of 
measure is set to inches, the parameter defaults to .394.}

\item{blank_margins}{When this option is TRUE, \strong{rptr} will use blank 
spaces and blank rows to create left and top margins, rather than rely 
on the editor to set margins.  When used, editor margins
should be set to zero.  Valid values are TRUE and FALSE. Default is
FALSE.  This option is only valid for \code{output_type = 'text'}.}
}
\value{
The report_spec with margins set as desired.
}
\description{
Sets the page margins for the report.  The units for this 
parameter can be inches or centimeters, depending on the units of measure 
specified on the \code{\link{create_report}} function.
}
\details{
The margins set with \code{set_margins} will be used for the entire report.  
Units for the margins
are specified by the \code{units} parameter on the 
\code{\link{create_report}} function.  Available units are 'inches' and 'cm'.
When the unit of measure is inches, default margins are 1 inch on the left and 
right, and .5 inches on top and bottom.  When the unit of measure is 
centimeters, default margins are 2.54 cm on left and right, and 1.27 cm
on top and bottom.

The \code{min_margin} parameter is used to set the minimum margin allowed
by the printer.  This value will be subtracted from the margin settings 
when the \code{blank_margins} option is used.  

Note that when using output type of text, and not using the 
\code{blank_margins} option, setting the margins only reduces
the area available for content on a page.  You must still set the actual
margins on the available editor to match those specified in 
\code{set_margins}.  Any mismatch may result in content not fitting properly
on the page. For best results, set the right and bottom margins to zero 
to allow for any overflow in spacing.
}
\examples{
library(rptr)
library(magrittr)

# Create a temporary file
tmp <- file.path(tempdir(), "mtcars.txt")

# Define report
rpt <- create_report(tmp, orientation="portrait") \%>\%
  titles("MTCARS Report") \%>\% 
  set_margins(top = 1, bottom =1) \%>\% 
  add_content(create_table(mtcars)) 

# Write the report
write_report(rpt)

}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
