% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{set_margins}
\alias{set_margins}
\title{Set page margins.}
\usage{
set_margins(
  x,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  min_margin = NULL
)
}
\arguments{
\item{x}{The report spec object.}

\item{top}{The top margin.}

\item{bottom}{The bottom margin.}

\item{left}{The left margin.}

\item{right}{The right margin.}

\item{min_margin}{The printer minimum margin.}
}
\value{
The report_spec with margins set as desired.
}
\description{
Set the page margins on the report spec object.
}
\details{
The margins will be used for the entire report.  Units for the margins
are inches.  The default margins are 1 inch on the left and right, and
.5 inch on the top and bottom.
}
\examples{
# Here is an example
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{titles}()},
\code{\link{write_report}()}
}
\concept{report}
