% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{set_margins}
\alias{set_margins}
\title{Set page margins.}
\usage{
set_margins(
  x,
  margin_top = 0.5,
  margin_bottom = 0.5,
  margin_left = 1,
  margin_right = 1
)
}
\arguments{
\item{x}{The report spec object.}

\item{margin_top}{The top margin.}

\item{margin_bottom}{The bottom margin.}

\item{margin_left}{The left margin.}

\item{margin_right}{The right margin.}
}
\value{
The report_spec with margins set as desired.
}
\description{
Set the page margins on the report spec object.
}
\details{
The margins will be used for the entire report.  Units for the margins
are inches.  The default margins are 1 inch on the left and right, and
.5 inch on the top and bottom.
}
\examples{
# Here is an example
}
