% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{spanning_header}
\alias{spanning_header}
\title{Defines a spanning header}
\usage{
spanning_header(
  x,
  span_cols,
  label = "",
  label_align = "center",
  level = 1,
  n = NULL
)
}
\arguments{
\item{x}{The table spec.}

\item{span_cols}{The columns to span.}

\item{label}{The label to apply to the spanning header.}

\item{label_align}{The alignment to use for the label.}

\item{level}{The level to use for the spanning header.  The bottom level is 
level 1, next level above is level 2, and so on.  By default, the level
is set to 1.}

\item{n}{The n value to use for the n label on the spanning header.}
}
\description{
Defines a spanning header
}
