% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.R
\name{stub}
\alias{stub}
\title{Defines a report stub}
\usage{
stub(
  x,
  vars,
  label = "",
  label_align = NULL,
  align = "left",
  width = NULL,
  format = NULL
)
}
\arguments{
\item{x}{The table spec.}

\item{vars}{A vector of quoted variable names from which to create the stub.}

\item{label}{The label for the report stub.}

\item{label_align}{The alignment for the label.  Valid values are 'left', 
'right', 'center', and 'centre'.  Default is 'left'.}

\item{align}{How to align the stub column.  Valid values are 'left', 
'right', 'center', and 'centre'.  Default is 'left'.}

\item{width}{The width of the stub, in report units of measure.}

\item{format}{A format to apply to the stub column.}
}
\value{
The modified table spec.
}
\description{
Combine one or more columns into a nested report stub.
}
\details{
Here are some details.
}
