% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_spec.r
\name{table_options}
\alias{table_options}
\title{Defines options for the table}
\usage{
table_options(x, first_row_blank = FALSE)
}
\arguments{
\item{x}{The table spec.}

\item{first_row_blank}{Whether to create a blank on the first row after the
table header.}
}
\description{
Defines options for the table
}
