% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{titles}
\alias{titles}
\title{Add titles to the report}
\usage{
titles(x, ..., location = "header", align = "center")
}
\arguments{
\item{x}{The report specification object.}

\item{...}{A set of title strings.}

\item{location}{The location to place the titles in the report.  Valid}

\item{align}{The position to align the titles.  Valid values are: "left"}
}
\description{
This function adds one or more to the report.  The titles will be added to
the page template, and thus appear on each page of the report.
}
\details{
The titles function accepts a set of strings of the desired title text.
The titles may be aligned center, left or right using the align parameter.
The titles may be located in the header, page body, or table according to
the location parameter.
}
\examples{
# Here is a comment
}
