% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{titles}
\alias{titles}
\title{Add titles to the report}
\usage{
titles(x, ..., align = "center", blank_row = "below")
}
\arguments{
\item{x}{The object to assign titles to.  Valid objects are a report,
a table, or a text specification.}

\item{...}{A set of title strings.}

\item{align}{The position to align the titles.  Valid values are 'left', 
'right', 'center' or 'centre'.}

\item{blank_row}{Where to place a blank row.  Valid values are 'above',
'below', 'both', or 'none'.}
}
\value{
The modified report.
}
\description{
This function adds one or more to the object.  If added to a report, 
the titles will be added to
the page template, and thus appear on each page of the report.
}
\details{
The titles function accepts a set of strings of the desired title text.
The titles may be aligned center, left or right using the align parameter.
The titles may be assigned to the report, a table or a text specification.
}
\examples{
# Here is a comment
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{set_margins}()},
\code{\link{write_report}()}
}
\concept{report}
