% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_registration_file}
\alias{write_registration_file}
\title{Write a registration file}
\usage{
write_registration_file(file_path)
}
\arguments{
\item{file_path}{The full or relative file name and path to create.}
}
\description{
This function will create a registration file to help determine
the correct \code{cpi} and \code{lpi} for editor/printer.  
The \code{cpi} and \code{lpi} are 
used in \code{output_type = "text"} to determine available space on
the page.
}
