% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{write_report}
\alias{write_report}
\title{Write a report to the file system}
\usage{
write_report(x, ...)
}
\arguments{
\item{x}{The report_spec object to write.}

\item{...}{Any parameters passed to the output-specific write functions.}
}
\value{
The report spec.
}
\description{
This function writes a report_spec object to the file system, using the
parameters provided in the object.
}
