% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_spec.R
\name{write_report}
\alias{write_report}
\title{Write a report to the file system}
\usage{
write_report(x)
}
\arguments{
\item{x}{The report object to write.}
}
\value{
The report spec, with settings modified during writing.
}
\description{
This function writes a report_spec object to the file system, using the
parameters provided in the object.
}
\details{
The function renders the report in the requested format, and writes it
to the location specified in the report \code{file_path} parameter.
}
\seealso{
Other report: 
\code{\link{add_content}()},
\code{\link{create_report}()},
\code{\link{footnotes}()},
\code{\link{options_fixed}()},
\code{\link{page_footer}()},
\code{\link{page_header}()},
\code{\link{set_margins}()},
\code{\link{titles}()}
}
\concept{report}
