% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officerx.r
\name{write_report}
\alias{write_report}
\title{Write a report to the file system}
\usage{
write_report(x, graphic_type = "png")
}
\arguments{
\item{x}{The report_spec object to write.}

\item{graphic_type}{If the report contains graphic content, this parameter
is used to specify the graphic format.  Value values are "png" and "svg".
The default is "png".}
}
\value{
The report spec.
}
\description{
This function writes a report_spec object to the file system, using the
parameters provided in the object.
}
